chore: Storybook에서 Next Image unoptimized 속성 추가
