+++ apps/workshop/package.json
@@ -30,6 +30,7 @@
     "@types/react-dom": "^18.0.11",
     "@vitejs/plugin-react-swc": "^3.0.0",
     "framer-motion": "^10.16.3",
+    "next": "^13.4.19",
     "prop-types": "^15.8.1",
     "storybook": "^7.0.23",
     "typescript": "^5.0.2",
